pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "Scriptum-Report"
version = "1.2.1"
description = "Generate DOCX and PPTX reports from templates using python-docx and python-pptx."
readme = "README.md"
requires-python = ">=3.10"
license = "PolyForm-Noncommercial-1.0.0 OR LicenseRef-SCRIPTUM-Commercial"
license-files = ["LICENSE.md", "LICENSES/*"]
authors = [
  { name = "Thomas Emmel", email = "temmel007@gmail.com" }
]
dependencies = [
  "python-docx>=1.2.00",
  "python-pptx>=1.0.00",
  "Pillow>=10.0.0"
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Intended Audience :: Developers",
  "Topic :: Office/Business",
  "Topic :: Software Development :: Libraries"
]

[tool.setuptools]
script-files = [ "scripts/convert_video.sh", "scripts/convert_video.py", "scripts/check_docx.py", 
                 "scripts/check_pptx.py", "scripts/check_rdf.py" ]

[tool.setuptools.packages.find]
include = ["Scriptum*", "CHANGELOG", "docs"]
exclude = ["tests**","tests*","*tests*"]

